add mock for leaflet, fix Map component
